Add files via upload
Problem -2 
Write a program to insert and remove the elements in a queue and display the final elements in the queue. 
Queue 
A queue is a particular kind of abstract data type or collection in which the entities in the collection are kept in order and the principal (or only) operations on the collection are the addition of entities to the rear terminal position, known as enqueue, and removal of entities from the front terminal position, known as dequeue. 
INPUT  
Format First line contains size of the queue. 
Second line contains the list of integers currently in the queue e.g 1 3 5 4
 Third line contains an integer 0 or 1 - 0 means dequeue, 1 means enqueue. If the third line contains 1, fourth line contains the integer to be enqueued. 
Note: First integer in the list of integers is the HEAD of the queue and the last integer is the TAIL of the queue. 
OUTPUT 
Print the modified queue after doing enqueue or dequeue operation. 
Sample Input 
4 1 3 5 4 1 6 
Sample Output 
1 3